+++ src/dyscord/client/__init__.py
@@ -1,9 +1,10 @@
-from .enumerations import INTENTS
+from .enumerations import INTENTS, DISCORD_EVENTS
 from .discord_client import DiscordClient
 from .api import API
 
 __all__ = [
     'INTENTS',
+    'DISCORD_EVENTS',
     'DiscordClient',
     'API',
 ]