+++ src/dyscord/client/discord_client.py
@@ -1,7 +1,6 @@
 import asyncio
 import functools
 import inspect
-import nest_asyncio  # type: ignore
 import platform
 import random
 import sys
@@ -12,10 +11,11 @@ from collections import defaultdict
 from pprint import pprint
 from typing import Any, Callable, Optional, List
 
-import websockets
+import nest_asyncio  # type: ignore
 import orjson as json
+import websockets
 
-from . import api, INTENTS
+from . import api, INTENTS, DISCORD_EVENTS
 
 from .. import utilities
 from .. import objects
@@ -46,7 +46,6 @@ class DiscordClient:
     application_id: Optional[str]
     intent: int
     ready: bool
-    cache: 'utilities.Cache'
 
     def __init__(self, token: str, application_id: Optional[str] = None):
         '''Instantiate a DiscordClient.
@@ -56,11 +55,10 @@ class DiscordClient:
             application_id (str): The application id. Can be left to None if client will not use Interactions.
         '''
         # Discord attributes
-        DiscordClient.token = token
-        DiscordClient.application_id = application_id
-        DiscordClient.intent = 0
-        DiscordClient.ready = False
-        DiscordClient.cache = utilities.Cache()
+        self.__class__.token = token
+        self.__class__.application_id = application_id
+        self.__class__.intent = 0
+        self.__class__.ready = False
 
         # Private attributes
         self._heartbeat_task = None
@@ -400,13 +398,13 @@ class DiscordClient:
 
         if event_type == 'READY':
             obj = objects.Ready().from_dict(data['d'])
-            DiscordClient.session_id = obj.session_id
-            DiscordClient.ready = True
-            DiscordClient.me = obj.user
+            self.__class__.session_id = obj.session_id
+            self.__class__.ready = True
+            self.__class__.me = obj.user
             self._log.info('Discord connection complete, we are ready!')
             self._log.info(f'We are now {self.me}')
 
-        elif not DiscordClient.ready:
+        elif not self.__class__.ready:
             self._log.info(f'Got event of type [{event_type}] before we were ready!')
             return
 
@@ -516,7 +514,6 @@ class DiscordClient:
             warnings.warn(f'Encountered unhandled event {event_type}')
 
         elif event_type == 'THREAD_CREATE':
-            pprint(data)
             obj = objects.ChannelImporter().from_dict(data['d'])
 
         elif event_type == 'THREAD_DELETE':
@@ -559,13 +556,15 @@ class DiscordClient:
         # Call own event handlers first.
         if hasattr(self, event_handler_name):
             self_function = getattr(self, event_handler_name)
-            await self_function(obj, data)
+            arg_len = len(inspect.signature(self_function).parameters)
+            arguments = (obj, data)
+            await self_function(*arguments[:arg_len])
 
         # Call user wrapped classes, functions and cotoutines.
         # TODO: Should we invoke a create_task when able to avoid blocking calls?
         arguments = (obj, data, self)
         if obj is not None:
-            for user_function in DiscordClient._wrapper_registrations[event_type]:
+            for user_function in self.__class__._wrapper_registrations[event_type]:
                 arg_len = len(inspect.signature(user_function).parameters)
                 assert arg_len >= 0 and arg_len <= 3
                 if asyncio.iscoroutinefunction(user_function):
@@ -573,7 +572,7 @@ class DiscordClient:
                 else:
                     user_function(*arguments[:arg_len])
 
-            for user_class in DiscordClient._wrapper_class_registrations:
+            for user_class in self.__class__._wrapper_class_registrations:
                 if hasattr(user_class, event_handler_name):
                     user_function = getattr(user_class, event_handler_name)
                     arg_len = len(inspect.signature(user_function).parameters)
@@ -586,7 +585,7 @@ class DiscordClient:
                         user_function(user_class, *arguments[:arg_len])
 
         # Handle the special case of the ANY event.
-        for user_function in DiscordClient._wrapper_registrations['ANY']:
+        for user_function in self.__class__._wrapper_registrations['ANY']:
             arg_len = len(inspect.signature(user_function).parameters)
             assert arg_len >= 0 and arg_len <= 3
             if asyncio.iscoroutinefunction(user_function):
@@ -594,7 +593,7 @@ class DiscordClient:
             else:
                 user_function(*arguments[:arg_len])
 
-        for user_class in DiscordClient._wrapper_class_registrations:
+        for user_class in self.__class__._wrapper_class_registrations:
             if hasattr(user_class, 'on_any'):
                 user_function = getattr(user_class, event_handler_name)
                 arg_len = len(inspect.signature(user_function).parameters)
@@ -677,8 +676,8 @@ class DiscordClient:
 
         The `client` is the DiscordClient instance.
         '''
-        if not hasattr(objects.DISCORD_EVENTS, event) and event != 'ANY':
-            raise ValueError(f'Attempted to bind to unknown event \'{event}\', must be exact match for existing {objects.DISCORD_EVENTS} entry.')
+        if not hasattr(DISCORD_EVENTS, event) and event != 'ANY':
+            raise ValueError(f'Attempted to bind to unknown event \'{event}\', must be exact match for existing {DISCORD_EVENTS} entry.')
 
         def func_wrapper(func):
             arg_len = len(inspect.signature(func).parameters)