+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 This project follows [Semantic Versioning](https://semver.org/) and [Keep A Change Log](https://keepachangelog.com/en/1.0.0/).
 
-## [v0.4.0] WIP
+## [v0.4.0] 2021-10-27
 ### Added
 - New set_all_intents() function in DiscordClient. Sets all intents to True.
 - Better example code.
@@ -10,10 +10,13 @@ This project follows [Semantic Versioning](https://semver.org/) and [Keep A Chan
 - Role now has all attributes set to `None` by default.
 - Begin restructure of documentation to make some sense.
 - Registering to a guild allows you to register different commands to different guilds.
+- TTL Cache to API.get_user, API.get_guild and API.get_channel.
 
 ### Changed
 - `InteractionStructure` renamed to `Interaction`.
 - `InteractionDataStructure` renamed to `InteractionData`.
+- `register_handler` renamed to `decorate_handler`.
+- `DISCORD_EVENTS` moved from `dyscord.objects` to `discord.client`.
 - Reworked the repo to be a bit cleaner.
 - Moved Dockerfiles to their own folder, with a 4x speedup on initial build times.
 - Attributes of classes updated to be `None` by default.
@@ -24,6 +27,10 @@ This project follows [Semantic Versioning](https://semver.org/) and [Keep A Chan
     - Channel (and subclasses).
     - Guild
     - Embed
+    - Message
+    - Ready
+    - User
+    - Member
 
 ### Fixed
 - Exposed the `__version__` string at the top module level.